scheduler/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
pyTelegramBotAPI==4.10.0
requests==2.28.2
telebot==0.0.5
urllib3==1.26.15
psycopg2-binary==2.9.3
alembic==1.8.1
SQLAlchemy==1.4.40
tenacity==8.0.1
pydantic==1.10.2
prettytable==3.4.1
click==8.1.3
beautifulsoup4==4.11.1
matplotlib==3.7.1
